data structures,greedy,*1100

+++ Codeforces Round #805 (Div. 3)/C. Train and Queries.cpp
@@ -70,7 +70,7 @@ void solve(void){
     	if(v[i] <= v[0]) continue;
     	debug(v);
 
-    	v[0] = ceil((v[0] + v[i]) / 2);
+    	v[0] = ceil(((float)v[0] + v[i]) / 2);
     }
     debug(v);
     nl;
